Move select function to seperate file
